+++ src/modules/scenarioeditor/ScenarioEditorMap.vue
@@ -114,9 +114,9 @@ watch(
   }),
   ({ engine, ...settings }) => {
     engine?.map.setViewConstraints({
-      extent: settings.extent ?? null,
-      minZoom: settings.minZoom ?? null,
-      maxZoom: settings.maxZoom ?? null,
+      extent: settings.extent,
+      minZoom: settings.minZoom,
+      maxZoom: settings.maxZoom,
     });
   },
   { immediate: true },